requirements.txt
python-dotenv
fastapi>=0.101.0
uvicorn[standard]==0.23.1
google-cloud-documentai
requests==2.31.0
google-genai